archive/requirements.txt
openai>=1.0.0
python-dotenv>=0.21.0
requests>=2.28.1
tqdm>=4.64.1
nltk>=3.7
beautifulsoup4>=4.11.1
pytest>=7.2.0 